Abstracts PWM timer functionality to a library.
TIMER1 is currently used to drive LED dimming via PWM. Prior to this commit,
it was coupled to the light module, and thus could not be used for other
purposes.

This commit decouples that module, allowing the timer to be used
to generate an additional 1Hz signal (which will be useful in future
IR communications procedures).